Addition of identity-file argument to init-ssh!
This allows a specific key to be nominated for the Git SSH connection.

An example of where this can be useful:
Consider a setup where keys A & B both authenticate to GitHub but with
different access to private repositories repo-A & repo-B.

Currently the first key that successfully authenticates to Github is used for the remainder of the Git
operation. So without a means of selecting the key to use one of the
private repos will be unavailable (git error: "Repository: Not Found.")

NOTE: JSCH doesn't use the ~/.ssh/config file where this would commonly be set.