Clarify usage of ssh config in README
The old indentation makes the impression that the `:ssh` key should live next to `:git`, not inside it. That took me a few hours to figure out.

Another issue: Using the `:identity-file` flag does not work as advertised. It has to be located in the folder returned by the jsch `getIdentityRepository` to be used.